requirements.txt
flask==2.3.2
werkzeug==2.3.6
flask-cors==3.0.10
pymongo==4.3.3
flask-swagger-ui==3.36.0
python-dotenv==0.21.0
marshmallow==3.19.0